Make MemAllocs into an abstract datatype.
All operations that traverse a MemAllocs log are now implemented
within the Lang.Crucible.LLVM.MemModel.MemLog module.
